refactor(desktop): address FF review — mvvm hook location, testSetup, async config read
- Move useResetTimeRangeOnGraphRework (+ its test) from src/renderer/hooks to src/mvvm/hooks
  per the MVVM architecture guidance; App.tsx imports it via the LLD/ alias.
- Rewrite the hook test to use renderHook from tests/testSetup with a real store (matching the
  desktop hook-test convention) instead of @testing-library/react.
- warnOnConfigMismatch: read .env files with the async fs.promises.readFile instead of readFileSync.

LIVE-31639
